+++ changelog.md
@@ -11,6 +11,7 @@
 7. Improved announcement, application detail, downloads, event, poll dialog UI
 8. Added import announcement from TruckersMP news
 9. Added calendar event counter
+10. Added non-member connect account reminder
 
 ## v3.4.2